Handle already existing file in target with same name With this the PatternFileMover is able to run properly even if there is a same-naming file in the target directory. Before this change PatternFileMover simply does nothing if it stumples upon this case.
Attention: PatternFileMover will remove the already existing file without any warning! This will be improved in a further version by adding options to choose from what to do in this case.

Furthermore PatternFileMover will not stop working if a name association is broken. It skips the broken association from now on.

+++ PatternFileMover/Form1.cs
@@ -91,6 +91,30 @@ namespace PatternFileMover
                 {
                     if (dataGridView1.Rows[i].Cells[0].Value.ToString().Contains(data.SearchPattern))
                     {
+                        if (File.Exists(
+                                data.TargetDirectory + 
+                                Path.DirectorySeparatorChar + 
+                                Path.GetFileName(dataGridView1.Rows[i].Cells[0].Value.ToString())
+                            )
+                        )
+                        {
+                            // delete the existing file before it gets replaced with the current
+                            // processed file
+                            File.Delete(
+                                data.TargetDirectory +
+                                Path.DirectorySeparatorChar +
+                                Path.GetFileName(dataGridView1.Rows[i].Cells[0].Value.ToString())
+                            );
+                        }
+
+                        if (!Directory.Exists(data.TargetDirectory + Path.DirectorySeparatorChar))
+                        {
+                            // the target directory does not existing
+                            // maybe a broken name association or a network drive is not available
+                            // skip this and go on
+                            continue;
+                        }
+
                         File.Move(
                             dataGridView1.Rows[i].Cells[0].Value.ToString(),
                             data.TargetDirectory + Path.DirectorySeparatorChar + Path.GetFileName(dataGridView1.Rows[i].Cells[0].Value.ToString())
